extract base scraper w/o browser
