+++ src/constants.js
@@ -9,12 +9,14 @@ export const DOLLAR_CURRENCY_SYMBOL = '$';
 export const DOLLAR_CURRENCY = 'USD';
 
 export const SCRAPE_PROGRESS_TYPES = {
+  INITIALIZING: 'INITIALIZING',
   START_SCRAPING: 'START_SCRAPING',
   LOGGING_IN: 'LOGGING_IN',
   LOGIN_SUCCESS: 'LOGIN_SUCCESS',
   LOGIN_FAILED: 'LOGIN_FAILED',
   CHANGE_PASSWORD: 'CHANGE_PASSWORD',
   END_SCRAPING: 'END_SCRAPING',
+  TERMINATING: 'TERMINATING',
 };
 
 export const LOGIN_RESULT = {
