+++ src/scrapers/base-scraper-with-browser.js
@@ -1,11 +1,9 @@
-import { EventEmitter } from 'events';
 import puppeteer from 'puppeteer';
 
+import { BaseScraper } from './base-scraper';
 import { SCRAPE_PROGRESS_TYPES, LOGIN_RESULT, GENERAL_ERROR } from '../constants';
-import { waitForNavigation, getCurrentUrl, NAVIGATION_ERRORS } from '../helpers/navigation';
+import { waitForNavigation, getCurrentUrl } from '../helpers/navigation';
 import { waitUntilElementFound, fillInput, clickButton } from '../helpers/elements-interactions';
-
-const SCRAPE_PROGRESS = 'SCRAPE_PROGRESS';
 
 function getKeyByValue(object, value) {
   return Object.keys(object).find(key => object[key] === value);
@@ -33,72 +31,23 @@ function handleLoginResult(scraper, loginResult) {
   }
 }
 
-function createErrorResult(errorType, errorMessage) {
+function createGeneralError() {
   return {
     success: false,
-    errorType,
-    errorMessage,
+    errorType: GENERAL_ERROR,
   };
 }
 
-function createTimeoutError(errorMessage) {
-  return createErrorResult(NAVIGATION_ERRORS.TIMEOUT, errorMessage);
-}
-
-function createGenericNavigationError(errorMessage) {
-  return createErrorResult(NAVIGATION_ERRORS.GENERIC, errorMessage);
-}
-
-function createGeneralError() {
-  return createErrorResult(GENERAL_ERROR);
-}
-
-class BaseScraperWithBrowser {
-  constructor(options) {
-    this.options = options;
-    this.eventEmitter = new EventEmitter();
-  }
-
+class BaseScraperWithBrowser extends BaseScraper {
   async initialize() {
+    this.emitProgress(SCRAPE_PROGRESS_TYPES.INITIALIZING);
+
     let env = null;
     if (this.options.verbose) {
       env = Object.assign({ DEBUG: '*' }, process.env);
     }
     this.browser = await puppeteer.launch({ env });
     this.page = await this.browser.newPage();
-  }
-
-  async scrape(credentials) {
-    this.emitProgress(SCRAPE_PROGRESS_TYPES.START_SCRAPING);
-    await this.initialize();
-
-    let loginResult;
-    try {
-      loginResult = await this.login(credentials);
-    } catch (e) {
-      loginResult = e.timeout ?
-        createTimeoutError(e.message) :
-        createGenericNavigationError(e.message);
-    }
-
-    let scrapeResult;
-    if (loginResult.success) {
-      try {
-        scrapeResult = await this.fetchData();
-      } catch (e) {
-        scrapeResult =
-          e.timeout ?
-            createTimeoutError(e.message) :
-            createGenericNavigationError(e.message);
-      }
-    } else {
-      scrapeResult = loginResult;
-    }
-
-    await this.terminate();
-    this.emitProgress(SCRAPE_PROGRESS_TYPES.END_SCRAPING);
-
-    return scrapeResult;
   }
 
   getLoginOptions() {
@@ -144,24 +93,9 @@ class BaseScraperWithBrowser {
     return handleLoginResult(this, loginResult);
   }
 
-  async fetchData() {
-    throw new Error(`fetchData() is not created in ${this.options.companyId}`);
-  }
-
   async terminate() {
+    this.emitProgress(SCRAPE_PROGRESS_TYPES.TERMINATING);
     await this.browser.close();
-  }
-
-  emitProgress(type) {
-    this.emit(SCRAPE_PROGRESS, { type });
-  }
-
-  emit(eventName, payload) {
-    this.eventEmitter.emit(eventName, this.options.companyId, payload);
-  }
-
-  onProgress(func) {
-    this.eventEmitter.on(SCRAPE_PROGRESS, func);
   }
 }
